# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Bitaqty Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vendors/issues": {
            "get": {
                "tags": [
                    "VendorIssue"
                ],
                "summary": "Issues",
                "description": "Vendor Issues",
                "operationId": "VendorIssues",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": ""
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "solution": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "is_solved": {
                                                        "type": "boolean",
                                                        "example": ""
                                                    },
                                                    "is_duplicate": {
                                                        "type": "boolean",
                                                        "example": ""
                                                    },
                                                    "issue_type": {
                                                        "type": "object",
                                                        "example": {
                                                            "id": 1,
                                                            "name": ""
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "VendorIssue"
                ],
                "summary": "Create Vendor Issue",
                "description": "Create Vendor Issue",
                "operationId": "VendorCreateIssue",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "issue_type_id",
                                    "description"
                                ],
                                "properties": {
                                    "issue_type_id": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/vendors/news": {
            "get": {
                "tags": [
                    "VendorNews"
                ],
                "summary": "News",
                "description": "Vendor News",
                "operationId": "VendorNews",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": ""
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "action": {
                                                        "type": "string",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}